+++ src/main/java/com/memesphere/domain/detail/dto/response/PriceInfoResponse.java
@@ -13,26 +13,26 @@ public class PriceInfoResponse {
     private Long coinId;
 
     @Schema(description = "현재가", example = "0.20")
-    private BigDecimal price;
+    private Double price;
 
     @Schema(description = "가격 변화량", example = "-0.03")
-    private BigDecimal priceChange;
+    private Double priceChange;
 
     @Schema(description = "가격 변화량 절대값", example = "0.03")
-    private BigDecimal priceChangeAbsolute;
+    private Double priceChangeAbsolute;
 
     @Schema(description = "가격 변화 방향", example = "down")
     private String priceChangeDirection;
 
     @Schema(description = "가격 변화율", example = "-6.35")
-    private BigDecimal priceChangeRate;
+    private Double priceChangeRate;
 
     @Schema(description = "거래량 가중 평균 가격", example = "-942.38")
-    private BigDecimal weightedAveragePrice;
+    private Double weightedAveragePrice;
 
     @Schema(description = "24h 최고가", example = "2500")
-    private BigDecimal highPrice;
+    private Double highPrice;
 
     @Schema(description = "24h 최저가", example = "1500")
-    private BigDecimal lowPrice;
+    private Double lowPrice;
 }
